Create new unit tests for the User schema and update file structure.
